+++ src/jvmMain/kotlin/com/sunnychung/application/multiplatform/giantlogviewer/ux/App.kt
@@ -12,7 +12,6 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.text.BasicText
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -34,12 +33,12 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import com.sunnychung.application.giantlogviewer.generated.resources.Res
+import com.sunnychung.application.giantlogviewer.generated.resources.fast_forward
+import com.sunnychung.application.giantlogviewer.generated.resources.fast_forward_filled
 import com.sunnychung.application.giantlogviewer.generated.resources.help
 import com.sunnychung.application.giantlogviewer.generated.resources.info
-import com.sunnychung.application.multiplatform.giantlogviewer.document.lightColorTheme
 import com.sunnychung.application.giantlogviewer.generated.resources.setting
 import com.sunnychung.application.multiplatform.giantlogviewer.document.ThemeDI
-import com.sunnychung.application.multiplatform.giantlogviewer.document.darkColorTheme
 import com.sunnychung.application.multiplatform.giantlogviewer.document.toColorTheme
 import com.sunnychung.application.multiplatform.giantlogviewer.extension.subscribeStateToEntity
 import com.sunnychung.application.multiplatform.giantlogviewer.io.GiantFileTextPager
@@ -65,10 +64,16 @@ fun App() {
         .subscribeStateToEntity(ThemeDI)
         .themes
 
+    var selectedFilePath by remember { mutableStateOf("") }
+    var fileViewState: FileViewState by remember(selectedFilePath) { mutableStateOf(FileViewState(File(selectedFilePath))) }
+
     print("App recompose - $themePreference")
 
     CompositionLocalProvider(LocalColor provides themePreference.toColorTheme()) {
         val colors = LocalColor.current
+
+        val viewerFocusRequester = remember { FocusRequester() }
+
         Column(Modifier.fillMaxSize()) {
             Row(
                 modifier = Modifier.fillMaxWidth().height(30.dp)
@@ -96,6 +101,18 @@ fun App() {
                     overflow = TextOverflow.Ellipsis,
                     modifier = Modifier.weight(1f).padding(horizontal = 6.dp),
                 )
+                if (fileViewState.isFileExist) {
+                    AppImage(
+                        resource = if (!fileViewState.isFollowing) Res.drawable.fast_forward else Res.drawable.fast_forward_filled,
+                        size = 20.dp,
+                        color = if (!fileViewState.isFollowing) colors.menuBarIconColor else colors.menuBarIconActivated,
+                        modifier = Modifier.padding(5.dp)
+                            .clickable {
+                                fileViewState.isFollowing = !fileViewState.isFollowing
+                                viewerFocusRequester.requestFocus()
+                            }
+                    )
+                }
                 AppImage(
                     resource = Res.drawable.help,
                     size = 20.dp,
@@ -116,7 +133,14 @@ fun App() {
                 )
             }
 
-            AppMainContent(onSelectFile = { selectedFileName = it?.name ?: "" })
+            AppMainContent(
+                fileViewState = fileViewState,
+                onSelectFile = {
+                    selectedFileName = it?.name ?: ""
+                    selectedFilePath = it?.path ?: ""
+                },
+                modifier = Modifier.focusRequester(viewerFocusRequester)
+            )
 
             HelpWindow(isVisible = isShowHelpWindow, onClose = { isShowHelpWindow = false })
             AboutWindow(isVisible = isShowAboutWindow, onClose = { isShowAboutWindow = false })
@@ -127,15 +151,15 @@ fun App() {
 
 @Composable
 @OptIn(ExperimentalComposeUiApi::class)
-private fun AppMainContent(modifier: Modifier = Modifier, onSelectFile: (File?) -> Unit) {
+private fun AppMainContent(modifier: Modifier = Modifier, fileViewState: FileViewState, onSelectFile: (File?) -> Unit) {
     val colors = LocalColor.current
+
+    val isNavigationLocked = fileViewState.isFollowing
 
     var selectedFilePath by remember { mutableStateOf("") }
 
     val viewerFocusRequester = remember { FocusRequester() }
     var filePager: GiantFileTextPager? by remember { mutableStateOf(null) }
-
-    var fileViewState by remember(selectedFilePath) { mutableStateOf(FileViewState(File(selectedFilePath))) }
 
     var isSearchBarVisible by remember { mutableStateOf(false) }
     var searchEntry by remember { mutableStateOf("") }
@@ -188,6 +212,8 @@ private fun AppMainContent(modifier: Modifier = Modifier, onSelectFile: (File?) 
 
                             println("f: ${uri.scheme} ${File(uri).absolutePath}")
                             selectedFilePath = File(uri).absolutePath
+
+                            viewerFocusRequester.requestFocus()
                         }
                     }
                 )
@@ -252,15 +278,20 @@ private fun AppMainContent(modifier: Modifier = Modifier, onSelectFile: (File?) 
                     SearchResultType.SomeResult
                 },
                 onToggleRegex = {
+                    if (isNavigationLocked) return@TextSearchBar
                     searchOptions = searchOptions.copy(isRegex = it)
                 },
                 onToggleCaseSensitive = {
+                    if (isNavigationLocked) return@TextSearchBar
                     searchOptions = searchOptions.copy(isCaseSensitive = it)
                 },
                 onToggleWholeWord = {
+                    if (isNavigationLocked) return@TextSearchBar
                     searchOptions = searchOptions.copy(isWholeWord = it)
                 },
                 onClickPrev = {
+                    if (isNavigationLocked) return@TextSearchBar
+
                     val regex = currentSearchRegex() ?: return@TextSearchBar
                     val pager = filePager ?: return@TextSearchBar
                     val result = try {
@@ -282,6 +313,8 @@ private fun AppMainContent(modifier: Modifier = Modifier, onSelectFile: (File?) 
                     searchEntryOfResult = searchEntry
                 },
                 onClickNext = {
+                    if (isNavigationLocked) return@TextSearchBar
+
                     val regex = currentSearchRegex() ?: return@TextSearchBar
                     val pager = filePager ?: return@TextSearchBar
                     if (pager.viewportStartBytePosition < fileViewState.fileLength) {