+++ src/jvmMain/kotlin/com/sunnychung/application/multiplatform/giantlogviewer/viewstate/FileViewState.kt
@@ -5,7 +5,11 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import java.io.File
 
-class FileViewState(file: File) {
+class FileViewState(val file: File) {
+
+    val isFileExist = file.isFile
 
     var fileLength by mutableStateOf(file.length())
+
+    var isFollowing by mutableStateOf(false)
 }